Fix non-working downloads per source setting
I was attempting to increase the number of downloads per source and
realized that no matter what I did I couldn't get it to retain the
setting. After some digging through the code I discovered that it was
missing in the Settings conversions. The property was missing from both
the `SaveSettingsAsync` and `GetFromEditableSettings` conversion
methods.